0.1.8-SNAPSHOT - fixed a bug on both renderers. Updated TextureRenderer with new scratchBuffer logic

+++ src/main/scala/s2d/shapes/BasicRenderer.scala
@@ -46,7 +46,7 @@ object BasicRenderer:
   private var colorLocation: Int = -1
   private var isInitialized: Boolean = false
   private val scratchBuffer: Ptr[GLfloat] =
-    stdlib.malloc(sizeof[GLfloat] * 2048.toULong).asInstanceOf[Ptr[GLfloat]]
+    stdlib.malloc(sizeof[GLfloat] * 2048.toUInt).asInstanceOf[Ptr[GLfloat]]
 
   def initialize(): Boolean =
     if isInitialized then return true
